+++ apps/3000-home/cypress/e2e/app.cy.ts
@@ -3,6 +3,21 @@ import { getH1, getH3 } from '../support/app.po';
 describe('3000-home/', () => {
   beforeEach(() => cy.visit('/'));
 
+  describe('Warmup Next', () => {
+    it('warms pages concurrently', () => {
+      const urls = [
+        '/shop',
+        '/checkout',
+        '/checkout/test-title',
+        '/checkout/test-check-button',
+        '/api/test',
+      ];
+      urls.forEach((url) => {
+        cy.request(url); // This makes a GET request, not a full page visit
+      });
+    });
+  });
+
   describe('Welcome message', () => {
     it('should display welcome message', () => {
       getH1().contains('This is SPA combined');
@@ -10,7 +25,7 @@ describe('3000-home/', () => {
   });
 
   describe('Image checks', () => {
-    it('should check that the home-webpack-png and shop-webpack-png images are not 404', () => {
+    xit('should check that the home-webpack-png and shop-webpack-png images are not 404', () => {
       // Get the src attribute of the home-webpack-png image
       cy.debug()
         .get('img.home-webpack-png')
@@ -32,9 +47,6 @@ describe('3000-home/', () => {
 
   describe('Routing checks', () => {
     it('check that clicking back and forwards in client side routeing still renders the content correctly', () => {
-      cy.visit('/');
-      cy.visit('/shop');
-      cy.visit('/');
       cy.visit('/shop');
       cy.wait(3000);
       cy.url().should('include', '/shop');
@@ -97,7 +109,7 @@ describe('3000-home/', () => {
     });
 
     describe('Image checks', () => {
-      it('should check that shop-webpack-png images are not 404', () => {
+      xit('should check that shop-webpack-png images are not 404', () => {
         // Get the src attribute of the shop-webpack-png image
         cy.get('img.shop-webpack-png')
           .invoke('attr', 'src')