+++ packages/nextjs-mf/src/plugins/NextFederationPlugin/index.ts
@@ -142,7 +142,6 @@ export class NextFederationPlugin {
     compiler: Compiler,
     isServer: boolean,
   ): ModuleFederationPluginOptions {
-    console.log(this._extraOptions.skipSharingNextInternals);
     const defaultShared = this._extraOptions.skipSharingNextInternals
       ? {}
       : retrieveDefaultShared(isServer);