+++ packages/node/src/plugins/NodeFederationPlugin.ts
@@ -12,6 +12,7 @@ import EntryChunkTrackerPlugin from './EntryChunkTrackerPlugin';
 interface NodeFederationOptions extends ModuleFederationPluginOptions {
   experiments?: Record<string, unknown>;
   debug?: boolean;
+  useRuntimePlugin?: boolean;
 }
 
 /**
@@ -31,6 +32,7 @@ class NodeFederationPlugin {
   private _options: ModuleFederationPluginOptions;
   private context: Context;
   private experiments: NodeFederationOptions['experiments'];
+  private useRuntimePlugin?: boolean;
 
   /**
    * Create a NodeFederationPlugin.
@@ -39,12 +41,13 @@ class NodeFederationPlugin {
    * @param {Context} context - The context for the NodeFederationPlugin
    */
   constructor(
-    { experiments, debug, ...options }: NodeFederationOptions,
+    { experiments, debug, useRuntimePlugin, ...options }: NodeFederationOptions,
     context: Context,
   ) {
     this._options = options || ({} as ModuleFederationPluginOptions);
     this.context = context || ({} as Context);
     this.experiments = experiments || {};
+    this.useRuntimePlugin = useRuntimePlugin || false;
   }
 
   /**
@@ -65,19 +68,21 @@ class NodeFederationPlugin {
   }
 
   private preparePluginOptions(): ModuleFederationPluginOptions {
+    this._options.runtimePlugins = [
+      ...(this.useRuntimePlugin ? [require.resolve('../runtimePlugin')] : []),
+      ...(this._options.runtimePlugins || []),
+    ];
+
     return {
       ...this._options,
       remotes: this._options.remotes || {},
+      runtimePlugins: this._options.runtimePlugins,
       // enable dts in browser by default
       dts: this._options.dts ?? false,
     };
   }
 
   private updateCompilerOptions(compiler: Compiler): void {
-    if (compiler.options && compiler.options.output) {
-      //todo, need to change / remove
-      compiler.options.output.importMetaName = 'remoteContainerRegistry';
-    }
     const chunkFileName = compiler.options?.output?.chunkFilename;
     const uniqueName =
       compiler?.options?.output?.uniqueName || this._options.name;