+++ packages/node/src/plugins/UniversalFederationPlugin.ts
@@ -18,6 +18,7 @@ interface NodeFederationOptions extends ModuleFederationPluginOptions {
   isServer: boolean;
   promiseBaseURI?: string;
   debug?: boolean;
+  useRuntimePlugin?: boolean;
 }
 
 /**
@@ -45,6 +46,37 @@ class UniversalFederationPlugin {
     this._options = options || ({} as NodeFederationOptions);
     this.context = context || ({} as NodeFederationContext);
     this.name = 'ModuleFederationPlugin';
+    if (this._options.useRuntimePlugin && this._options.isServer) {
+      this._options.runtimePlugins = this._options.runtimePlugins
+        ? this._options.runtimePlugins.concat([
+            require.resolve('../runtimePlugin.js'),
+          ])
+        : [require.resolve('../runtimePlugin.js')];
+    }
+  }
+
+  private updateCompilerOptions(compiler: Compiler): void {
+    compiler.options.output.chunkFormat = 'commonjs';
+    if (compiler.options.output.enabledLibraryTypes === undefined) {
+      compiler.options.output.enabledLibraryTypes = ['commonjs-module'];
+    } else {
+      compiler.options.output.enabledLibraryTypes.push('commonjs-module');
+    }
+
+    const chunkFileName = compiler.options?.output?.chunkFilename;
+    const uniqueName =
+      compiler?.options?.output?.uniqueName || this._options.name;
+    if (
+      typeof chunkFileName === 'string' &&
+      uniqueName &&
+      !chunkFileName.includes(uniqueName)
+    ) {
+      const suffix = `-[chunkhash].js`;
+      compiler.options.output.chunkFilename = chunkFileName.replace(
+        '.js',
+        suffix,
+      );
+    }
   }
 
   /**
@@ -52,7 +84,7 @@ class UniversalFederationPlugin {
    * @param {Compiler} compiler - The webpack compiler
    */
   apply(compiler: Compiler) {
-    const { isServer, debug, ...options } = this._options;
+    const { isServer, debug, useRuntimePlugin, ...options } = this._options;
     const { webpack } = compiler;
     if (!process.env['FEDERATION_WEBPACK_PATH']) {
       process.env['FEDERATION_WEBPACK_PATH'] = getWebpackPath(compiler);
@@ -63,8 +95,15 @@ class UniversalFederationPlugin {
       compiler.options.target === 'node' ||
       compiler.options.target === 'async-node'
     ) {
-      new NodeFederationPlugin(options, this.context).apply(compiler);
-      new StreamingTargetPlugin({ ...options, debug }).apply(compiler);
+      if (useRuntimePlugin) {
+        this.updateCompilerOptions(compiler);
+        new ModuleFederationPlugin({
+          ...options,
+        }).apply(compiler);
+      } else {
+        new NodeFederationPlugin(options, this.context).apply(compiler);
+        new StreamingTargetPlugin({ ...options, debug }).apply(compiler);
+      }
     } else {
       new ModuleFederationPlugin(options).apply(compiler);
     }