+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,11 @@
 # [3.1.0](https://github.com/module-federation/universe/compare/utils-3.0.2...utils-3.1.0) (2023-10-26)
 
+## 3.0.14
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.9
+
 ## 3.0.13
 
 ### Patch Changes